+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 - Added httpc call GET for redirect from Google
+- Change ct to clickTag
 
 ## [0.3.1] - 2018-04-21
 ### Added